add: bookmark_category 추가 / refact: Setter 제거, 파일 구조 변경
