--- src/main/java/com/umc/owncast/domain/Member/dto/MemberResponseDTO.java
@@ -1,11 +1,0 @@
-package com.umc.owncast.domain.Member.dto;
-
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class MemberResponseDTO {
-
-}
