--- src/main/java/com/umc/owncast/domain/Member/dto/MemberSignUpRequestDTO.java
@@ -1,25 +1,0 @@
-package com.umc.owncast.domain.Member.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Pattern;
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
-
-@Builder
-@Getter
-@Setter
-@Data
-public class MemberSignUpRequestDTO {
-
-    @Pattern(regexp = "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$", message = "이메일 형식이 올바르지 않습니다.")
-    @NotBlank(message = "이메일은 필수 입력 값입니다.")
-    private String email;
-
-    @Pattern(regexp = "^[a-zA-z0-9]{6,20}$", message = "비밀번호 형식이 올바르지 않습니다.")
-    private String password;
-
-    @NotBlank(message = "닉네임은 필수 입력 값입니다.")
-    private String nickname;
-}
