--- src/main/java/com/umc/owncast/domain/Member/entity/Member.java
@@ -1,69 +1,0 @@
-package com.umc.owncast.domain.Member.entity;
-
-import com.umc.owncast.common.entity.BaseTimeEntity;
-import com.umc.owncast.domain.Cast.entity.Cast;
-import com.umc.owncast.domain.Language.entity.Language;
-import com.umc.owncast.domain.Playlist.entity.Playlist;
-import com.umc.owncast.enums.Status;
-import com.umc.owncast.mapping.Bookmark;
-import com.umc.owncast.mapping.CastLike;
-import com.umc.owncast.mapping.CastSave;
-import com.umc.owncast.mapping.MemberPrefer;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@Entity
-@Table(name = "member")
-@AllArgsConstructor
-public class Member extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 50)
-    private String username;
-
-    @Column(nullable = false, length = 50)
-    private String password;
-
-    @Column(length = 50)
-    private String email;
-
-    private LocalDate inactiveDate;
-
-    @Enumerated(EnumType.STRING)
-    @Column(columnDefinition = "VARCHAR(255) DEFAULT 'ACTIVE'")
-    private Status status;
-
-    private boolean emailVerified;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "language_id")
-    private Language language;
-
-    /*@OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<MemberPrefer> memberPreferList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<Playlist> playlistList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<Cast> castList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<CastLike> castLikeList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<CastSave> castSaveList = new ArrayList<>();
-
-    @OneToMany(mappedBy = "member", cascade = CascadeType.ALL)
-    private List<Bookmark> bookmarkList = new ArrayList<>();*/
-}
