--- src/main/java/com/umc/owncast/domain/Member/repository/MemberRepository.java
@@ -1,7 +1,0 @@
-package com.umc.owncast.domain.Member.repository;
-
-import com.umc.owncast.domain.Member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
