--- src/main/java/com/umc/owncast/domain/Member/service/UserService.java
@@ -1,11 +1,0 @@
-package com.umc.owncast.domain.Member.service;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@RequiredArgsConstructor
-@Service
-public class UserService {
-}
