--- src/main/java/com/umc/owncast/domain/Sentence/entity/Sentence.java
@@ -1,40 +1,0 @@
-package com.umc.owncast.domain.Sentence.entity;
-
-import com.umc.owncast.common.entity.BaseTimeEntity;
-import com.umc.owncast.domain.Cast.entity.Cast;
-import com.umc.owncast.mapping.Bookmark;
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.awt.print.Book;
-import java.util.ArrayList;
-import java.util.List;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@Entity
-@Table(name = "maincategory")
-@AllArgsConstructor
-public class Sentence extends BaseTimeEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Integer index;
-
-    @Column(nullable = false)
-    private String originalSentence;
-
-    @Column(nullable = false)
-    private String translatedSentence;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "cast_id")
-    private Cast cast;
-
-    /*@OneToMany(mappedBy = "sentence", cascade = CascadeType.ALL)
-    private List<Bookmark> bookmarkList = new ArrayList<>();*/
-}
