+++ src/main/java/com/umc/owncast/domain/bookmark_category/entity/BookmarkCategory.java
@@ -1,0 +1,28 @@
+package com.umc.owncast.domain.bookmark_category.entity;
+
+import com.umc.owncast.common.entity.BaseTimeEntity;
+import com.umc.owncast.domain.member.entity.Member;
+import com.umc.owncast.domain.playlist.entity.Playlist;
+import com.umc.owncast.domain.sentence.entity.Sentence;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@Entity
+@Table(name = "bookmark_category")
+@AllArgsConstructor
+public class BookmarkCategory extends BaseTimeEntity{
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "sentence_id")
+    private Sentence sentence;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "playlist_id")
+    private Playlist playlist;
+}
