+++ src/main/java/com/umc/owncast/domain/cast_category/entity/CastCategory.java
@@ -1,0 +1,27 @@
+package com.umc.owncast.domain.cast_category.entity;
+
+import com.umc.owncast.common.entity.BaseTimeEntity;
+import com.umc.owncast.domain.cast.entity.Cast;
+import com.umc.owncast.domain.category.sub_category.entity.SubCategory;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@Entity
+@Table(name = "maincategory")
+@AllArgsConstructor
+public class CastCategory extends BaseTimeEntity{
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "cast_id")
+    private Cast cast;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "category_id")
+    private SubCategory subCategory;
+}
