+++ src/main/java/com/umc/owncast/domain/cast_like/entity/CastLike.java
@@ -1,0 +1,27 @@
+package com.umc.owncast.domain.cast_like.entity;
+
+import com.umc.owncast.common.entity.BaseTimeEntity;
+import com.umc.owncast.domain.cast.entity.Cast;
+import com.umc.owncast.domain.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.*;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@Entity
+@Table(name = "maincategory")
+@AllArgsConstructor
+public class CastLike extends BaseTimeEntity{
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "cast_id")
+    private Cast cast;
+}
