+++ src/main/java/com/umc/owncast/domain/enums/Status.java
@@ -1,0 +1,5 @@
+package com.umc.owncast.domain.enums;
+
+public enum Status {
+    ACTIVE, INACTIVE
+}
