+++ src/main/java/com/umc/owncast/domain/enums/Voice.java
@@ -1,0 +1,5 @@
+package com.umc.owncast.domain.enums;
+
+public enum Voice {
+    OFFICIAL, CASUAL
+}
