+++ src/main/java/com/umc/owncast/domain/member/dto/MemberLoginRequestDTO.java
@@ -1,0 +1,23 @@
+package com.umc.owncast.domain.member.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Pattern;
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+
+@Getter
+@Builder
+public class MemberLoginRequestDTO {
+
+    @NotEmpty(message = "이메일은 필수 입력값입니다.")
+    @Pattern(regexp = "^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+.[A-Za-z]{2,6}$", message = "이메일 형식에 맞지 않습니다.")
+    private String email;
+
+    @NotEmpty(message = "비밀번호는 필수 입력값입니다.")
+    private String password;
+
+    public UsernamePasswordAuthenticationToken toAuthentication() {
+        return new UsernamePasswordAuthenticationToken(email, password);
+    }
+}
