+++ src/main/java/com/umc/owncast/domain/member/dto/MemberResponseDTO.java
@@ -1,0 +1,11 @@
+package com.umc.owncast.domain.member.dto;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class MemberResponseDTO {
+
+}
