+++ src/main/java/com/umc/owncast/domain/member/repository/MemberRepository.java
@@ -1,0 +1,7 @@
+package com.umc.owncast.domain.member.repository;
+
+import com.umc.owncast.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
