--- src/main/java/com/umc/owncast/enums/Status.java
@@ -1,5 +1,0 @@
-package com.umc.owncast.enums;
-
-public enum Status {
-    ACTIVE, INACTIVE
-}
