--- src/main/java/com/umc/owncast/enums/Voice.java
@@ -1,5 +1,0 @@
-package com.umc.owncast.enums;
-
-public enum Voice {
-    OFFICIAL, CASUAL
-}
