--- src/main/java/com/umc/owncast/mapping/CastSave.java
@@ -1,28 +1,0 @@
-package com.umc.owncast.mapping;
-
-import com.umc.owncast.common.entity.BaseTimeEntity;
-import com.umc.owncast.domain.Cast.entity.Cast;
-import com.umc.owncast.domain.Member.entity.Member;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@Entity
-@Table(name = "maincategory")
-@AllArgsConstructor
-public class CastSave extends BaseTimeEntity{
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "cast_id")
-    private Cast cast;
-}
